requirements.txt
click==8.1.3
dominate==2.5.2
Flask~=2.1.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.39
visitor==0.1.3
Werkzeug~=2.2.1
WTForms==2.3.3
gunicorn==20.1.0
python-dotenv~=0.20.0
psycopg2==2.93
psycopg2-binary==2.9.3